requirements.txt
# ─── Core Data Pipeline ───
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
psycopg2-binary>=2.9

# ─── Configuration ───
python-dotenv>=1.0

# ─── Map Matching (OSRM client) ───
requests>=2.31

# ─── FastAPI Backend ───
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
asyncpg>=0.30.0
pydantic-settings>=2.0

# ─── Validation (optional — for DBCV metric) ───
# hdbscan>=0.8.33

# ─── Data Download ───
beautifulsoup4>=4.12
